+++ src/Transformers/LinearDiscriminantAnalysis.php
@@ -4,6 +4,7 @@ namespace Rubix\ML\Transformers;
 
 use Tensor\Matrix;
 use Rubix\ML\DataType;
+use Rubix\ML\Persistable;
 use Rubix\ML\Datasets\Labeled;
 use Rubix\ML\Datasets\Dataset;
 use Rubix\ML\Specifications\SamplesAreCompatibleWithTransformer;
@@ -26,7 +27,7 @@ use const Rubix\ML\EPSILON;
  * @package     Rubix/ML
  * @author      Andrew DalPino
  */
-class LinearDiscriminantAnalysis implements Transformer, Stateful
+class LinearDiscriminantAnalysis implements Transformer, Stateful, Persistable
 {
     /**
      * The target number of dimensions to project onto.