+++ src/Transformers/TfIdfTransformer.php
@@ -3,6 +3,7 @@
 namespace Rubix\ML\Transformers;
 
 use Rubix\ML\DataType;
+use Rubix\ML\Persistable;
 use Rubix\ML\Datasets\Dataset;
 use Rubix\ML\Specifications\SamplesAreCompatibleWithTransformer;
 use Rubix\ML\Exceptions\InvalidArgumentException;
@@ -29,7 +30,7 @@ use function is_null;
  * @package     Rubix/ML
  * @author      Andrew DalPino
  */
-class TfIdfTransformer implements Transformer, Stateful, Elastic
+class TfIdfTransformer implements Transformer, Stateful, Elastic, Persistable
 {
     /**
      * The amount of additive (Laplace) smoothing to add to the inverse document