+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - 0.3.0
     - Now throws Rubix\ML namespaced exceptions
+    - Stateful Transformers are now Persistable
     - Added K Best feature selector
     - Added Sentence tokenizer
     - Now can silence Verbose learners with silence() method